Restructure routine generators
Move routine generators into separate folder.
Delete redundant App.DTO project.
Update routine generators to comply with new data schema

+++ BLL.App.DTO/FiveThreeOneWendlerMaxesDto.cs
@@ -2,9 +2,9 @@ namespace App.DTO
 {
     public class FiveThreeOneWendlerMaxesDto
     {
-        public decimal SquatMax { get; set; }
-        public decimal DeadliftMax { get; set; }
-        public decimal BenchPressMax { get; set; }
-        public decimal OverHeadPressMax { get; set; }
+        public float SquatMax { get; set; }
+        public float DeadliftMax { get; set; }
+        public float BenchPressMax { get; set; }
+        public float OverHeadPressMax { get; set; }
     }
 }
